requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pandas==2.3.2
numpy==2.2.6
yfinance==0.2.65
python-dotenv==1.1.1
